+++ frontend/src/libs/map-sdk/types/adapter.types.ts
